pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trnsci"
version = "0.1.0"
description = "Scientific computing suite for AWS Trainium via NKI (FFT, BLAS, RAND, SOLVER, SPARSE, TENSOR)"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Scott Friedman", email = "scott@playgroundlogic.co" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = []

[project.optional-dependencies]
fft = ["trnfft>=0.7"]
blas = ["trnblas>=0.3"]
rand = ["trnrand>=0.1"]
solver = ["trnsolver>=0.1"]
sparse = ["trnsparse>=0.1"]
tensor = ["trntensor>=0.1"]
all = [
    "trnfft>=0.7",
    "trnblas>=0.3",
    "trnrand>=0.1",
    "trnsolver>=0.1",
    "trnsparse>=0.1",
    "trntensor>=0.1",
]
neuron = [
    # neuronxcc is not on public PyPI; ships with the Deep Learning AMI Neuron.
    # Pinned >=2.24 because all sub-project NKI kernels use the 2.24+
    # nc_matmul calling convention.
    "neuronxcc>=2.24",
    "torch-neuronx>=2.9",
]
dev = [
    "pytest>=7.0",
    "pytest-benchmark>=4.0",
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocs-monorepo-plugin>=1.1",
]

[project.urls]
Homepage = "https://github.com/trnsci/trnsci"
Documentation = "https://trnsci.github.io/trnsci/"
Repository = "https://github.com/trnsci/trnsci"
Issues = "https://github.com/trnsci/trnsci/issues"

[tool.setuptools]
# Meta-package: no Python code of its own.
py-modules = []

[tool.pytest.ini_options]
testpaths = ["tests"]
